requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# OpenAI and Azure integration
openai==1.75.0
semantic-kernel==1.28.1
azure-identity==1.15.0
azure-search-documents==11.4.0

# Web scraping and caching
requests==2.31.0
beautifulsoup4==4.12.0
lxml==5.1.0
cachetools==5.3.0

# Template rendering
jinja2==3.1.2

# Testing
pytest==7.4.4 